reelrungen2/requirements.txt
moviepy==1.0.3
Pillow>=9.5.0,<10.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
edge-tts>=6.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
